2.1.3 Bug Fix & Help Panel
I found the root cause of the bug mentioned in the last commit. I used
the id of pomodorino twice. That has been fixed.

I removed the help icon from the settings menu and now have it on the
main page. This way, there will also be information about how to use the
timer and manipulate its settings.