2.3.1 Tooltip
Added a help tooltip to the "Strict Mode" option, so users understand
the feature.

Removed the "save button" from the settings menu and replaced it with a
message informing users that changes will be saved automatically. This
will be done by saving on the modal close (Foundation fires an event).